@@ -1,40 +1,34 @@
-import {
-  TextField,
-  Grid,
-  Typography,
-} from "@mui/material";
-import React from "react";
+import React, { useMemo } from 'react';
+import { useTranslation } from 'react-i18next';
 
-function TextInput({ text, setText, setDataChanged }) {
+import { Grid, TextField } from '@mui/material';
+
+const TextInput = ({ text, onChangeData }) => {
+  const { t } = useTranslation();
+
+  const value = useMemo(() => text ?? '', [text]);
+
   return (
-    <div align="center">
-      <Grid
-        container
-        direction={"column"}
-        align="left"
-        columns={{ xs: 12, sm: 12, md: 12, lg: 12, xl: 12 }}
-      >
-        <Grid item sx={{ pb: 2 }}>
-          <Typography variant="p1"> Type Answer here:</Typography>
-        </Grid>
-        <Grid item sx={{ pb: 2 }}>
-          <TextField
-            value={text}
-            placeholder="Enter Answer"
-            label="Answer"
-            name="quiz text answer"
-            variant="outlined"
-            onChange={(t) => {
-              {
-                setDataChanged(true);
-                setText(t.target.value);
-              }
-            }}
-          />
-        </Grid>
+    <Grid container>
+      <Grid item>
+        <TextField
+          value={value}
+          placeholder={t('Enter Answer')}
+          label={t('Answer')}
+          variant="outlined"
+          onChange={(t) => onChangeData(t.target.value)}
+        />
       </Grid>
-    </div>
+    </Grid>
   );
-}
+};
+
+TextInput.handleSave = ({ saveFn, data, id, type }) => {
+  saveFn({
+    id,
+    data,
+    type,
+  });
+};
 
 export default TextInput;
