@@ -1,52 +1,53 @@
-import { APP_DATA_TYPES, QUESTION_TYPES } from "../components/constants/constants";
+import { APP_SETTING_NAMES, QUESTION_TYPES } from '../config/constants';
 
 const buildDatabase = (appContext) => ({
-  appData: [
+  appSettings: [
     {
+      id: 'question-list',
       data: {
-        list: ["id4", "id5", "id6"],
+        list: ['id6', 'id5', 'id4'],
       },
-      type: APP_DATA_TYPES.QUESTION_LIST,
+      name: APP_SETTING_NAMES.QUESTION_LIST,
     },
     {
-      id: "id4",
+      id: 'id4',
       data: {
-        question: "What is the capital of France?",
-        questionType: QUESTION_TYPES.MULTIPLE_CHOICE,
+        question: 'What is the capital of France?',
+        type: QUESTION_TYPES.MULTIPLE_CHOICES,
         choices: [
-          { choice: "London", isCorrect: false },
-          { choice: "Paris", isCorrect: true },
-          { choice: "New York", isCorrect: false },
-          { choice: "Tokyo", isCorrect: false },
+          { choice: 'London', isCorrect: false },
+          { choice: 'Paris', isCorrect: true },
+          { choice: 'New York', isCorrect: false },
+          { choice: 'Tokyo', isCorrect: false },
         ],
       },
-      type: APP_DATA_TYPES.QUESTION,
+      name: APP_SETTING_NAMES.QUESTION,
     },
     {
-      id: "id5",
+      id: 'id5',
       data: {
-        question: "How happy are you?",
-        questionType: QUESTION_TYPES.SLIDER,
-        leftText: "Sad",
-        rightText: "Happy",
-        correctValue: 20,
+        question: 'How happy are you?',
+        type: QUESTION_TYPES.SLIDER,
+        min: 10,
+        max: 90,
+        value: 20,
       },
-      type: APP_DATA_TYPES.QUESTION,
+      name: APP_SETTING_NAMES.QUESTION,
     },
     {
-      id: "id6",
+      id: 'id6',
       data: {
-        question: "What is a baby cat called?",
-        questionType: QUESTION_TYPES.TEXT_INPUT,
-        answer: "kitten",
+        question: 'What is a baby cat called?',
+        type: QUESTION_TYPES.TEXT_INPUT,
+        answer: 'kitten',
       },
-      type: APP_DATA_TYPES.QUESTION,
+      name: APP_SETTING_NAMES.QUESTION,
     },
   ],
   members: [
     {
       id: appContext.memberId,
-      name: "mock-member",
+      name: 'mock-member',
     },
   ],
 });
