@@ -1,11 +1,11 @@
-import { createTheme } from "@mui/material/styles";
+import { createTheme } from '@mui/material/styles';
 
 const graaspTheme = createTheme({
-    palette: {
-      primary: {
-        main: "#555BD9",
-      },
+  palette: {
+    primary: {
+      main: '#555BD9',
     },
-  });
+  },
+});
 
-export default graaspTheme
+export default graaspTheme;
